Abstract registry compatibility selector into a separate component
